# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ATM",
    "version": "v1"
  },
  "paths": {
    "/ClientById": {
      "get": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ClientByUsername": {
      "get": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "Username",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CreateClientAsync": {
      "post": {
        "tags": [
          "Atm"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/TransactionHistoryAsync": {
      "get": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "IdAccount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/InsertMoneyAsync": {
      "post": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "IdAccount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WithdrawMoneyAsync": {
      "post": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "IdAccount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/SendMoneyAsync": {
      "post": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "IdAccount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdRecipientAccount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "variableNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "note",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "noteForRecipient",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Authenticate": {
      "get": {
        "tags": [
          "Atm"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "pinCode",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/MonthlyInterest": {
      "get": {
        "tags": [
          "Atm"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "application/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "text/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CurrentAccount": {
        "type": "object",
        "properties": {
          "idAccount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "SavingsAccount": {
        "type": "object",
        "properties": {
          "idAccount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "interestPerMensem": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "Client": {
        "type": "object",
        "properties": {
          "idClient": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "currentAccount": {
            "$ref": "#/components/schemas/CurrentAccount"
          },
          "savingsAccount": {
            "$ref": "#/components/schemas/SavingsAccount"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "pinCodeHash": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}